--- src/components/range/test.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title></title>
-</head>
-<body>
-  <svg id="svg" width="500" height="80" style="margin-left: 20px"></svg>
-  <script type="text/javascript" src="slider.js"></script>
-</body>
-</html>
